패링 적용, 스태미너 및 체력 ui 적용

+++ Kronos/Assets/Scripts/PlayerStatus.cs
@@ -9,8 +9,8 @@ public static class PlayerStatus
 
     static public int HPMax;
     static public int HP;
-    static public int StaminaMax;
-    static public int Stamina;
+    static public float StaminaMax;
+    static public float Stamina;
 
     static public int attack;
     static public int shield;
